:recycle: Move Discord string utils into API
