+++ chatter-discord/src/main/java/me/axieum/mcmod/chatter/impl/discord/callback/minecraft/PlayerConnectionCallback.java
@@ -3,8 +3,8 @@ package me.axieum.mcmod.chatter.impl.discord.callback.minecraft;
 import me.axieum.mcmod.chatter.impl.discord.ChatterDiscord;
 import me.axieum.mcmod.chatter.impl.discord.config.module.MessageConfig.DimensionPredicate;
 import me.axieum.mcmod.chatter.impl.discord.util.DiscordDispatcher;
-import me.axieum.mcmod.chatter.impl.discord.util.StringUtils;
 import me.axieum.mcmod.chatter.impl.util.MessageFormat;
+import me.axieum.mcmod.chatter.impl.util.StringUtils;
 import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents.Disconnect;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents.Join;
