+++ chatter-discord/src/main/java/me/axieum/mcmod/chatter/impl/discord/callback/minecraft/ReceiveChatCallback.java
@@ -4,8 +4,9 @@ import me.axieum.mcmod.chatter.api.event.chat.ChatEvents;
 import me.axieum.mcmod.chatter.impl.discord.ChatterDiscord;
 import me.axieum.mcmod.chatter.impl.discord.config.module.MessageConfig.DimensionPredicate;
 import me.axieum.mcmod.chatter.impl.discord.util.DiscordDispatcher;
-import me.axieum.mcmod.chatter.impl.discord.util.StringUtils;
+import me.axieum.mcmod.chatter.impl.discord.util.FormatUtils;
 import me.axieum.mcmod.chatter.impl.util.MessageFormat;
+import me.axieum.mcmod.chatter.impl.util.StringUtils;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.Text;
 import org.jetbrains.annotations.Nullable;
@@ -21,7 +22,7 @@ public class ReceiveChatCallback implements ChatEvents.ReceiveChat
             final MessageFormat formatter = new MessageFormat()
                     .datetime("datetime")
                     .tokenize("player", player.getDisplayName().getString())
-                    .tokenize("message", StringUtils.minecraftToDiscord(raw))
+                    .tokenize("message", FormatUtils.minecraftToDiscord(raw))
                     .tokenize("world", StringUtils.getWorldName(player.world));
 
             // Dispatch a message to all configured channels
